streamlit-bakehouse/requirements.txt
databricks-sql-connector
databricks-sdk
streamlit
pandas
plotly
posit-sdk
